Add files via upload
A database has been created and configured based on PostgreSQL, an ORM model using SQLAlchemy, and a migration system using Alembic. For ease of installation, a `proj_setup.bat` file has been created, which allows you to install all the necessary dependencies right away and launches Docker and the database.